# openapi/emre-hub-ai-actions.yaml
openapi: 3.1.0
info:
  title: Emre Hub AI Actions
  version: 1.0.0
  description: |
    Safe, AI-ready HTTP API for Emre Hub (personal life dashboard).

    Designed for a future Custom GPT Action, OpenAI Assistants tool, or
    internal agent. Routes are **non-destructive** (no DELETE). Writes are
    inserts or upserts only and are always audit-logged server-side.

    Authentication: `Authorization: Bearer <AI_API_KEY>` (server env).
    All data access is scoped to a single hub owner UUID on the server.
  contact:
    name: Emre Hub
servers:
  - url: https://YOUR_DEPLOYMENT_HOST
    description: Production (replace with your deployed Next.js URL)
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/ai/habits/log:
    patch:
      operationId: upsertHabitLog
      tags: [Habits]
      summary: Mark habit complete / incomplete
      description: |
        Non-destructive upsert of a habit log for a date (defaults to today).
        Never deletes rows.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitLogInput"
            example:
              habit_id: "00000000-0000-0000-0000-000000000010"
              completed: true
              note: "Done after dinner"
      responses:
        "200":
          description: Upserted habit log
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/OkEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/HabitLog"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: |
        Shared secret from server env `AI_API_KEY`.
        Send as `Authorization: Bearer <AI_API_KEY>`.
        Fallback header also accepted by the server: `x-ai-api-key`.
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ValidationError:
      description: Request body or query failed Zod validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    OkEnvelope:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          const: true
        data: {}
    ErrorEnvelope:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          const: false
        error:
          type: string
        details: {}
    IsoDate:
      type: string
      pattern: '^\d{4}-\d{2}-\d{2}$'
      example: "2026-07-19"
...
    HabitLogInput:
      type: object
      required: [habit_id, completed]
      properties:
        habit_id: { type: string, minLength: 1 }
        log_date: { $ref: "#/components/schemas/IsoDate" }
        completed: { type: boolean }
        note: { type: string, maxLength: 500 }
    HabitLog:
      type: object
      properties:
        id: { type: string, format: uuid }
        habit_id: { type: string, format: uuid }
        log_date: { $ref: "#/components/schemas/IsoDate" }
        completed: { type: boolean }
        count: { type: integer }
        note: { type: string, nullable: true }
...